renamed c definition files
